[bugfix] Store packages’ eXist version requirement
This info is needed when clients request packages that are compatible with specific versions of eXist. Otherwise, all packages will be presumed compatible with all versions of eXist (e.g., >2.2.0), and the app is unable to tell clients which version of a package is compatible with their actual version of eXist.