[bugfix] Include old versions when searching pkgs
Clients request the newest version of a package compatible with their processor. Before this PR, eXist ignored previous versions of a package, so the repo would only supply the current package. As a result, users of older versions of eXist could be told there is no version of the package compatible with their system, when in fact older versions of the package compatible with their system should be supplied. This fixes the problem.