Update dependencies to templating v1.0.2
v1.0.1 added the lib:resolve-apps and lib:parse-param functions required in this PR, and v1.0.2 improved performance of lib:resolve-apps. (v1.0.3 is the current version in the templating repo but hasn't been published yet; once it is, we'll need to update our use of lib:parse-param to use the switch the param names from "start" to "delim-start" and "end" to "delim-end". In that respect, v1.0.3 perhaps should be reclassed as a breaking change?)